Move class 'physical_register'
