+++ libvm68kapi/bits/vm68k/register.h
@@ -21,128 +21,9 @@
 
 #include <vm68k/data>
 #include <bits/vm68kapi.h>
-#include <utility>
-#include <cstdint>
 
 namespace vm68k
 {
-    class _VM68KAPI_PUBLIC physical_register
-    {
-    public:
-        /**
-         * Use-counting pointers.
-         *
-         * This class should work as 'std::shared_ptr'.
-         */
-        class _VM68KAPI_PUBLIC pointer
-        {
-        private:
-            physical_register *_ptr = nullptr;
-
-        public:
-            pointer() noexcept = default;
-
-            explicit pointer(physical_register *x) noexcept
-            :
-                _ptr {x}
-            {
-                if (_ptr != nullptr) {
-                    (_ptr->_use_count)++;
-                }
-            }
-
-            pointer(const pointer &other) noexcept
-            :
-                pointer(other._ptr)
-            {
-                // Nothing to do.
-            }
-
-            pointer(pointer &&other) noexcept
-            {
-                swap(other);
-            }
-
-        public:
-            pointer &operator =(const pointer &other) noexcept
-            {
-                if (this != &other) {
-                    if (_ptr != nullptr) {
-                        (_ptr->_use_count)--;
-                    }
-                    _ptr = other._ptr;
-                    if (_ptr != nullptr) {
-                        (_ptr->_use_count)++;
-                    }
-                }
-                return *this;
-            }
-
-            pointer &operator =(pointer &&other) noexcept
-            {
-                swap(other);
-                return *this;
-            }
-
-        public:
-            ~pointer() noexcept
-            {
-                if (_ptr != nullptr) {
-                    (_ptr->_use_count)--;
-                }
-            }
-
-        public:
-            void swap(pointer &other) noexcept
-            {
-                std::swap(_ptr, other._ptr);
-            }
-
-        public:
-            physical_register *get() const noexcept
-            {
-                return _ptr;
-            }
-
-        public:
-            physical_register &operator *() const noexcept
-            {
-                return *_ptr;
-            }
-
-        public:
-            physical_register *operator ->() const noexcept
-            {
-                return _ptr;
-            }
-        };
-
-    private:
-        std::uint32_t _value = 0;
-
-        unsigned int _use_count = 0;
-
-    public:
-        physical_register() noexcept = default;
-
-        physical_register(const physical_register &) = delete;
-
-    public:
-        void operator =(const physical_register &) = delete;
-
-    public:
-        explicit constexpr operator std::uint32_t() const noexcept
-        {
-            return _value;
-        }
-
-    public:
-        constexpr unsigned int use_count() const
-        {
-            return _use_count;
-        }
-    };
-
     /**
      * Abstract class of register files.
      */
